Move root suspense boundary to logged in layout
